requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
python-dotenv
pyTelegramBotAPI
jinja2
python-multipart